# generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "janus-server",
    "description": "",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/operations/{id}": {
      "get": {
        "tags": [
          "operations"
        ],
        "operationId": "get_operation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Operation id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_OperationView"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataResponse_OperationView": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "kind",
              "status",
              "target_kind",
              "correlation_id",
              "version",
              "created_at",
              "updated_at"
            ],
            "properties": {
              "correlation_id": {
                "type": "string"
              },
              "created_at": {
                "type": "string"
              },
              "current_step": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "id": {
                "type": "string"
              },
              "kind": {
                "type": "string"
              },
              "problem": {},
              "progress": {},
              "result": {},
              "status": {
                "type": "string"
              },
              "target_id": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "target_kind": {
                "type": "string"
              },
              "updated_at": {
                "type": "string"
              },
              "version": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "current_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "detail": {
            "type": "string"
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}